1eVSOLID

+++ TP1/Forme.cpp
@@ -1,6 +1,6 @@
 #include "Forme.h"
 
-Forme::Forme(sf::Color _color, int _edgewidth)
+Forme::Forme(sf::Color _color, float _edgewidth)
 	:color(_color), edgewidth(_edgewidth)
 {
 }
@@ -10,7 +10,7 @@ sf::Color Forme::getColor()
 	return color;
 }
 
-int Forme::getEdgewitdh()
+float Forme::getEdgewitdh()
 {
 	return edgewidth;
 }
